requirements.txt
flask==3.0.3      
nltk==3.8.1      
numpy==1.24.4     
pandas==2.0.3      
spacy==3.7.4
pyspellchecker==0.8.1
# python -m spacy download en_core_web_sm
python-dateutil==2.8.2
pyTelegramBotAPI==4.13.0
telebot==0.0.5
Telethon==1.30.3
pytz==2023.3.post1
scikit-learn==1.3.0
APscheduler==3.10.4
scipy==1.10.1
# openpyxl==3.1.2
# streamlit==1.31.1
# openpyxl==3.1.2
# scipy==1.10.1
